Simplify Project by moving some functions to card utils
